Difficulty now increases with time and a few other minor things were added

+++ app/src/main/java/sgeorgiev/org/spaceshipgame/Projectile.java
@@ -62,7 +62,7 @@ public class Projectile implements GameObject {
         hitBox.bottom = y + bitmap.getHeight();
     }
 
-    public int getSpeed() {
+    public float getSpeed() {
         return speed;
     }
